==> src/downstream/dnanexus_notebooks/12c_phewas_map_samples_to_phecodes.ipynb <==
# install.packages(c("readr", "devtools"))
# devtools::install_github('PheWAS/PheWAS')


library(dplyr)
library(tidyr)
library(readr)
library(stringr)
library(PheWAS)


# Read files
icd_file <- "/mnt/project/notebooks/bmi/data/icd_raw.csv.gz"
icd_codes_file <- "/mnt/project/fields/data/phenotype_processing/coding19.tsv"
pheno_file <- "/mnt/project/notebooks/bmi/data/pheno.csv.gz"

# Prepare icd dataframe in long format
icd_df <- read_csv(icd_file, col_types = cols(sample_names = col_character()))
icd_df <- icd_df %>%
  mutate(icd = strsplit(icd, "\\|")) %>%
  unnest(icd)

# Prepare icd codes file to get the actual icd 10 codes
icd_codes_df <- read_tsv(icd_codes_file, col_types = cols(.default = col_character())) %>%
  mutate(coding = gsub(" ", "", coding),
         code = word(meaning, 1))

# Merge coding and meaning information, also add other required info
icd_df <- icd_df %>%
  left_join(icd_codes_df, by = c("icd" = "coding")) %>%
  mutate(count = 1, vocabulary_id = "ICD10", id = sample_names, code = as.character(code)) %>%
  select(id, vocabulary_id, code, index=count)


# Read pheno df and add sex info
pheno_df <- read_csv(pheno_file, col_types = cols(sample_names = col_character()))
id_sex <- pheno_df %>%
  select(id = sample_names, sex) %>%
  mutate(sex = ifelse(sex == "Male", "M", "F"))

# Only keep pheno df samples in icd df
icd_df <- icd_df %>%
  semi_join(id_sex, by = "id")


# Map ICD10 codes to phecodes
phe_df <- createPhenotypes(icd_df, min.code.count=1, id.sex=id_sex,
                           add.phecode.exclusions=T, translate=T, vocabulary.map=PheWAS::phecode_map_icd10)


length(phe_df)

head(phe_df)

head(phe_df)

# Remove any duplicate rows
phe_df <- phe_df[!duplicated(phe_df$id), ]


length(phe_df)

# Save results_df in CSV.gz format
write.csv(phe_df, gzfile("phecodes.csv.gz"), row.names = FALSE)


system('dx upload phecodes.csv.gz --path /notebooks/bmi/data/downstream/phecodes/')


dim(phe_df)


dim(id_sex)
