==> src/downstream/dnanexus_notebooks/14b_proteomics_bmi_assoc.ipynb <==
# !pip install statsmodels
# !pip install scikit-learn


import os
import dxpy
import numpy as np
import pandas as pd
import json
from scipy.stats import pearsonr, norm
import re
from functools import reduce

import statsmodels.api as sm
from sklearn.preprocessing import StandardScaler
from patsy import dmatrices

# Get associated proteins

assoc_prot_df = pd.DataFrame()

for gene in ["MC4R", "SLTM", "APBA1", "UBR3",  "BSN", "RIF1", "YLPM1", "PCSK1", "UBR2", "GIGYF1", "SLC5A3", "BLTP1", "GRM7"]:
    gene_df = pd.read_csv(
        f"/mnt/project/notebooks/bmi/data/downstream/proteomics/{gene}_protein_assoc.csv.gz"
    )
    gene_df = gene_df.loc[
        (gene_df.p_value<0.05/2923)&
        (gene_df.nobs>40000)
    ]
    gene_df["gene"] = gene
    assoc_prot_df = pd.concat((assoc_prot_df, gene_df))


gene_df = pd.read_csv(
    f"/mnt/project/notebooks/bmi/data/downstream/proteomics/PCSK1_protein_assoc.csv.gz"
)

gene_df.sort_values("p_value")

assoc_prot_df = assoc_prot_df.reset_index(drop=True)


assoc_prot_df

def upload_file_to_project(filename, proj_dir):
    dxpy.upload_local_file(filename, folder=proj_dir, parents=True)
    print(f"*********{filename} uploaded!!*********")
    os.remove(filename)
    return


assoc_prot_df.to_csv("./assoc_protein_gene_carrier.csv.gz", index=False)
proj_dir = "/notebooks/bmi/data/downstream/proteomics/"
upload_file_to_project("assoc_protein_gene_carrier.csv.gz", proj_dir)


# Read and parse NPX df

npx_df = pd.read_csv(
    "/mnt/project/notebooks/proteomics/data/npx_processed.csv.gz", 
    dtype={"sample_names": str},
    usecols=["sample_names"] + list(assoc_prot_df.protein.values)
)


def normalize(ser):
    scaler = StandardScaler()
    norm_ser = scaler.fit_transform(ser.to_frame())
    return pd.Series(norm_ser.flatten(), name=ser.name)


norm_npx_df = npx_df.set_index("sample_names").apply(normalize)
norm_npx_df["sample_names"] = npx_df.sample_names


# Read and process the pheno file

# get pheno file
pheno_df = pd.read_csv("/mnt/project/notebooks/bmi/data/pheno.csv.gz", dtype={"sample_names": str})


def normalize_covariates(pheno_df, covariates, exclude=set(["genetic_sex"])):
    norm_pheno_df = pheno_df.copy()
    for cov in covariates:
        if cov not in exclude:
            scaler = StandardScaler()
            norm_pheno_df[cov] = scaler.fit_transform(norm_pheno_df.loc[:, [cov]])
    return norm_pheno_df

def rint_normalization(ser):
    ranks = ser.rank()
    normalized = norm.ppf((ranks - 0.5)/ranks.notna().sum())
    return normalized

pheno_df["bmi_rint"] = pheno_df.groupby(["ancestry_pred", "sex"])["bmi"].transform(rint_normalization)
pheno_df["genetic_sex"] = (pheno_df.genetic_sex=="Female").astype(int)
pheno_df["age_2"] = pheno_df.age**2
pheno_df["age_sex"] = pheno_df.age*pheno_df.genetic_sex
pheno_df["exome_release_batch"] = (pheno_df.exome_release_batch=="50K Release").astype(int)


# normalized pheno df
numerical_covariates = ["age", "age_2", "age_sex", "bmi_prs"] + [f"genetic_pca{i}" for i in range(1, 11)]
categorical_covariates = ["exome_release_batch", "genetic_sex"]

norm_pheno_df = normalize_covariates(pheno_df, numerical_covariates)


protein_pheno_df = norm_pheno_df.merge(norm_npx_df, on="sample_names")

def train_model_sm(X, y, protein):
    model = sm.OLS(y, X)
    results = model.fit()
    r2 = results.rsquared
    coef = results.params.loc[protein]
    se = results.bse.loc[protein]
    stat = results.tvalues.loc[protein]
    conf = results.conf_int().loc[protein].values
    p_val = results.pvalues.loc[protein]
    return coef, se, conf[0], conf[1], p_val, stat, r2, results.nobs


def create_feature_label(gene_pheno_df, protein, numerical_covariates, categorical_covariates):
    equation = f"bmi_rint ~ {protein} + " + " + ".join(numerical_covariates) + " + " + " + ".join([f"C({cc})" for cc in categorical_covariates]) 
    y, X = dmatrices(equation, data=gene_pheno_df, return_type='dataframe')
    return X, y 


def train(gene_pheno_df, protein, numerical_covariates, categorical_covariates):
    X, y = create_feature_label(gene_pheno_df, protein, numerical_covariates, categorical_covariates)
    coef, se, ci_low, ci_high, p_val, stat, r2, nobs = train_model_sm(X, y, protein)
    return coef, se, ci_low, ci_high, p_val, stat, r2, nobs


def get_bmi_assoc(ser, pheno_df, numerical_covariates, categorical_covariates):
    gene_pheno_df = pheno_df.copy()
    protein = ser.protein
    coef, se, ci_low, ci_high, p_val, stat, r2, nobs = train(gene_pheno_df, protein, numerical_covariates, categorical_covariates)
    return pd.Series({
        "gene": ser.gene, "protein": protein, "coef": coef, "se": se, 
        "ci_low": ci_low, "ci_high": ci_high, "p_value": p_val, "stat": stat, 
        "r_squared": r2, "nobs": nobs
    })


bmi_assoc_df = assoc_prot_df.apply(
    get_bmi_assoc, 
    args=(protein_pheno_df, numerical_covariates, categorical_covariates,),
    axis=1
)

bmi_assoc_df

bmi_assoc_df.to_csv("./protein_bmi_assoc.csv.gz", index=False)
proj_dir = "/notebooks/bmi/data/downstream/proteomics/"
upload_file_to_project("protein_bmi_assoc.csv.gz", proj_dir)
